Add files via upload
Photography presentation website, that contains index, about us, portfolio, services presentation and contact pages. The contact part, uses a register/login as a way of contacting the page administrator. The registered user can communicate with the administrator through a chat section. The administrator can contact any user through that chat and can add photos to the portfolio and index pages.